+++ packages/core/vitest.config.ts
@@ -23,6 +23,7 @@ export default defineConfig({
       'src/features/layouts/findMatchingNavigationItem.spec.ts',
       'src/features/layouts/isFrameForAppUrl.spec.ts',
       'src/features/chromeActions/**/*.spec.ts',
+      'src/components/**/*.spec.ts',
       '../sdk/src/**/*.spec.ts',
     ],
   },